LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/toBlob-cross-realm-callback-report-exception-expected.txt

CONSOLE MESSAGE: Error: PASS


PASS toBlob() reports the exception from its callback in the callback's global object
